plugins/webdav: use readdir instead of readdir_r
glibc 2.24 started deprecating readdir_r. It says that glibc and
other implementantions are thread safe:

In the current POSIX.1 specification (POSIX.1-2008), readdir() is
not required to be thread-safe.  However, in modern implementations
(including the glibc implementation), concurrent calls to read‐
dir() that specify different directory streams are thread-safe.

So convert readdir_r users to readdir.